LPS-48992 Rename to IntervalAction since it's not related with indexer anymore (as it used to be in the past) and move it to another package.
